extract block overlay rendering from `GTClient` to a new dedicated class `BlockOverLayRenderer`
