+++ src/main/java/gregtech/common/GTClient.java
@@ -7,7 +7,6 @@ package gregtech.common;
 
 import static gregtech.api.enums.Mods.Forestry;
 import static gregtech.api.enums.Mods.GregTech;
-import static org.lwjgl.opengl.GL11.GL_LINE_LOOP;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -15,48 +14,29 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeSet;
-import java.util.function.Function;
 
 import net.minecraft.block.Block;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.entity.EntityClientPlayerMP;
-import net.minecraft.client.renderer.GLAllocation;
-import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.resources.IReloadableResourceManager;
 import net.minecraft.client.resources.IResourceManager;
 import net.minecraft.client.resources.IResourceManagerReloadListener;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.init.Blocks;
 import net.minecraft.item.ItemStack;
-import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumChatFormatting;
 import net.minecraft.util.MathHelper;
-import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.util.StatCollector;
 import net.minecraft.world.ChunkCoordIntPair;
-import net.minecraftforge.client.event.DrawBlockHighlightEvent;
 import net.minecraftforge.client.event.sound.SoundSetupEvent;
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.common.util.ForgeDirection;
 import net.minecraftforge.event.entity.living.LivingEvent;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.oredict.OreDictionary;
 
-import org.lwjgl.opengl.GL11;
-import org.lwjgl.opengl.GL20;
-
 import com.glodblock.github.nei.recipes.FluidRecipe;
 import com.glodblock.github.nei.recipes.extractor.GregTech5RecipeExtractor;
-import com.gtnewhorizon.structurelib.alignment.IAlignment;
-import com.gtnewhorizon.structurelib.alignment.IAlignmentProvider;
 
-import appeng.api.util.IOrientable;
-import appeng.tile.misc.TileInterface;
-import codechicken.lib.vec.Rotation;
-import codechicken.lib.vec.Scale;
-import codechicken.lib.vec.Transformation;
-import codechicken.lib.vec.Translation;
 import cpw.mods.fml.client.event.ConfigChangedEvent;
 import cpw.mods.fml.common.FMLCommonHandler;
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
@@ -67,7 +47,6 @@ import cpw.mods.fml.common.network.FMLNetworkEvent;
 import gregtech.api.GregTechAPI;
 import gregtech.api.covers.CoverRegistry;
 import gregtech.api.enums.GTValues;
-import gregtech.api.enums.ItemList;
 import gregtech.api.enums.Materials;
 import gregtech.api.gui.GUIColorOverride;
 import gregtech.api.gui.modularui.FallbackableSteamTexture;
@@ -76,15 +55,9 @@ import gregtech.api.hazards.HazardProtection;
 import gregtech.api.hazards.HazardProtectionTooltip;
 import gregtech.api.interfaces.IBlockOnWalkOver;
 import gregtech.api.interfaces.IToolStats;
-import gregtech.api.interfaces.tileentity.ICoverable;
-import gregtech.api.interfaces.tileentity.IGregTechTileEntity;
-import gregtech.api.interfaces.tileentity.ITurnable;
 import gregtech.api.items.MetaGeneratedItem;
 import gregtech.api.items.MetaGeneratedTool;
-import gregtech.api.metatileentity.BaseMetaPipeEntity;
-import gregtech.api.metatileentity.BaseMetaTileEntity;
 import gregtech.api.metatileentity.MetaPipeEntity;
-import gregtech.api.metatileentity.implementations.MTEBasicMachine;
 import gregtech.api.net.GTPacketClientPreference;
 import gregtech.api.recipe.RecipeCategory;
 import gregtech.api.util.ColorsMetadataSection;
@@ -94,10 +67,10 @@ import gregtech.api.util.GTModHandler;
 import gregtech.api.util.GTMusicSystem;
 import gregtech.api.util.GTPlayedSound;
 import gregtech.api.util.GTUtility;
+import gregtech.client.BlockOverlayRenderer;
 import gregtech.client.GTMouseEventHandler;
 import gregtech.client.SeekingOggCodec;
 import gregtech.client.capes.GTCapesLoader;
-import gregtech.common.blocks.BlockFrameBox;
 import gregtech.common.blocks.ItemMachines;
 import gregtech.common.config.Client;
 import gregtech.common.handlers.SprayColorInfiniteKeybindHandler;
@@ -120,7 +93,6 @@ import gregtech.loaders.ExtraIcons;
 import gregtech.loaders.misc.GTBees;
 import gregtech.loaders.preload.GTPreLoad;
 import gregtech.nei.NEIGTConfig;
-import ic2.api.tile.IWrenchable;
 import paulscode.sound.SoundSystemConfig;
 import paulscode.sound.SoundSystemException;
 
@@ -128,52 +100,6 @@ import paulscode.sound.SoundSystemException;
 // GTProxy
 
 public class GTClient extends GTProxy {
-
-    private static final List<Block> ROTATABLE_VANILLA_BLOCKS;
-
-    private static final int[][] GRID_SWITCH_TABLE = new int[][] { { 0, 5, 3, 1, 2, 4 }, { 5, 0, 1, 3, 2, 4 },
-        { 1, 3, 0, 5, 2, 4 }, { 3, 1, 5, 0, 2, 4 }, { 4, 2, 3, 1, 0, 5 }, { 2, 4, 3, 1, 5, 0 }, };
-
-    // don't ask. these "just works"
-    private static final Transformation ROTATION_MARKER_TRANSFORM_CENTER = new Scale(0.5);
-    private static final Transformation[] ROTATION_MARKER_TRANSFORMS_SIDES_TRANSFORMS = {
-        new Scale(0.25).with(new Translation(0, 0, 0.375))
-            .compile(),
-        new Scale(0.25).with(new Translation(0.375, 0, 0))
-            .compile(),
-        new Scale(0.25).with(new Translation(0, 0, -0.375))
-            .compile(),
-        new Scale(0.25).with(new Translation(-0.375, 0, 0))
-            .compile(), };
-    private static final int[] ROTATION_MARKER_TRANSFORMS_SIDES = { -1, -1, 2, 0, 3, 1, -1, -1, 0, 2, 3, 1, 0, 2, -1,
-        -1, 3, 1, 2, 0, -1, -1, 3, 1, 1, 3, 2, 0, -1, -1, 3, 1, 2, 0, -1, -1 };
-    private static final Transformation[] ROTATION_MARKER_TRANSFORMS_CORNER = {
-        new Scale(0.25).with(new Translation(0.375, 0, 0.375))
-            .compile(),
-        new Scale(0.25).with(new Translation(-0.375, 0, 0.375))
-            .compile(),
-        new Scale(0.25).with(new Translation(0.375, 0, -0.375))
-            .compile(),
-        new Scale(0.25).with(new Translation(-0.375, 0, -0.375))
-            .compile(), };
-    private static int rotationMarkerDisplayList;
-    private static boolean rotationMarkerDisplayListCompiled = false;
-
-    static {
-        ROTATABLE_VANILLA_BLOCKS = Arrays.asList(
-            Blocks.piston,
-            Blocks.sticky_piston,
-            Blocks.furnace,
-            Blocks.lit_furnace,
-            Blocks.dropper,
-            Blocks.dispenser,
-            Blocks.chest,
-            Blocks.trapped_chest,
-            Blocks.ender_chest,
-            Blocks.hopper,
-            Blocks.pumpkin,
-            Blocks.lit_pumpkin);
-    }
 
     public final PollutionRenderer mPollutionRenderer = new PollutionRenderer();
     private final List<Materials> mPosR;
@@ -201,7 +127,6 @@ public class GTClient extends GTProxy {
     private boolean mAnimationDirection;
     private GTClientPreference mPreference;
     private boolean mFirstTick = false;
-    private static final int ROTATION_MARKER_RESOLUTION = 120;
     private int mReloadCount;
     private float renderTickTime;
 
@@ -311,250 +236,6 @@ public class GTClient extends GTProxy {
         mMoltenNegB = Collections.singletonList(Materials.InfusedEntropy);
     }
 
-    private static void drawGrid(DrawBlockHighlightEvent aEvent, boolean showCoverConnections, boolean aIsWrench,
-        boolean aIsSneaking) {
-
-        GL11.glPushMatrix();
-
-        GL11.glEnable(GL11.GL_BLEND);
-        GL11.glBlendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA);
-        GL11.glDisable(GL11.GL_TEXTURE_2D);
-
-        // pause shader
-        int program = GL11.glGetInteger(GL20.GL_CURRENT_PROGRAM);
-        GL20.glUseProgram(0);
-
-        MovingObjectPosition target = aEvent.target;
-        EntityPlayer player = aEvent.player;
-        double camX = player.lastTickPosX + (player.posX - player.lastTickPosX) * (double) aEvent.partialTicks;
-        double camY = player.lastTickPosY + (player.posY - player.lastTickPosY) * (double) aEvent.partialTicks;
-        double camZ = player.lastTickPosZ + (player.posZ - player.lastTickPosZ) * (double) aEvent.partialTicks;
-        GL11.glTranslated(target.blockX - (int) camX, target.blockY - (int) camY, target.blockZ - (int) camZ);
-        GL11.glTranslated(0.5D - (camX - (int) camX), 0.5D - (camY - (int) camY), 0.5D - (camZ - (int) camZ));
-        final int tSideHit = target.sideHit;
-        Rotation.sideRotations[tSideHit].glApply();
-        // draw grid
-        GL11.glTranslated(0.0D, -0.502D, 0.0D);
-        GL11.glLineWidth(2.5F);
-        GL11.glColor4f(0.0F, 0.0F, 0.0F, 0.5F);
-        GL11.glBegin(GL11.GL_LINES);
-        GL11.glVertex3d(+.50D, .0D, -.25D);
-        GL11.glVertex3d(-.50D, .0D, -.25D);
-        GL11.glVertex3d(+.50D, .0D, +.25D);
-        GL11.glVertex3d(-.50D, .0D, +.25D);
-        GL11.glVertex3d(+.25D, .0D, -.50D);
-        GL11.glVertex3d(+.25D, .0D, +.50D);
-        GL11.glVertex3d(-.25D, .0D, -.50D);
-        GL11.glVertex3d(-.25D, .0D, +.50D);
-        final TileEntity tTile = player.worldObj.getTileEntity(target.blockX, target.blockY, target.blockZ);
-        final Block block = player.worldObj.getBlock(target.blockX, target.blockY, target.blockZ);
-        final int meta = player.worldObj.getBlockMetadata(target.blockX, target.blockY, target.blockZ);
-
-        // draw connection indicators
-        int tConnections = 0;
-        if (tTile instanceof ICoverable iCoverable) {
-            if (showCoverConnections) {
-                for (final ForgeDirection tSide : ForgeDirection.VALID_DIRECTIONS) {
-                    if (iCoverable.hasCoverAtSide(tSide)) tConnections |= tSide.flag;
-                }
-            } else if (tTile instanceof BaseMetaTileEntity baseMetaTile && baseMetaTile.getAlignment() == null) {
-                if (!aIsSneaking) tConnections |= baseMetaTile.getFrontFacing().flag;
-                else if (baseMetaTile.getMetaTileEntity() instanceof MTEBasicMachine basicMachine) {
-                    tConnections |= basicMachine.mMainFacing.flag;
-                }
-            } else if (tTile instanceof BaseMetaPipeEntity pipeEntity) tConnections = pipeEntity.mConnections;
-        } else if (tTile instanceof IWrenchable wrenchable) {
-            tConnections |= ForgeDirection.getOrientation(wrenchable.getFacing()).flag;
-        } else if (ROTATABLE_VANILLA_BLOCKS.contains(block)) {
-            tConnections |= ForgeDirection.getOrientation(meta).flag;
-        } else if (tTile instanceof TileInterface tileInterface) tConnections |= tileInterface.getUp()
-            .getOpposite().flag;
-
-        if (tConnections != 0) {
-            for (ForgeDirection tSide : ForgeDirection.VALID_DIRECTIONS) {
-                if ((tConnections & tSide.flag) != 0) {
-                    switch (GRID_SWITCH_TABLE[target.sideHit][tSide.ordinal()]) {
-                        case 0 -> {
-                            GL11.glVertex3d(+.25D, .0D, +.25D);
-                            GL11.glVertex3d(-.25D, .0D, -.25D);
-                            GL11.glVertex3d(-.25D, .0D, +.25D);
-                            GL11.glVertex3d(+.25D, .0D, -.25D);
-                        }
-                        case 1 -> {
-                            GL11.glVertex3d(-.25D, .0D, +.50D);
-                            GL11.glVertex3d(+.25D, .0D, +.25D);
-                            GL11.glVertex3d(-.25D, .0D, +.25D);
-                            GL11.glVertex3d(+.25D, .0D, +.50D);
-                        }
-                        case 2 -> {
-                            GL11.glVertex3d(-.50D, .0D, -.25D);
-                            GL11.glVertex3d(-.25D, .0D, +.25D);
-                            GL11.glVertex3d(-.50D, .0D, +.25D);
-                            GL11.glVertex3d(-.25D, .0D, -.25D);
-                        }
-                        case 3 -> {
-                            GL11.glVertex3d(-.25D, .0D, -.50D);
-                            GL11.glVertex3d(+.25D, .0D, -.25D);
-                            GL11.glVertex3d(-.25D, .0D, -.25D);
-                            GL11.glVertex3d(+.25D, .0D, -.50D);
-                        }
-                        case 4 -> {
-                            GL11.glVertex3d(+.50D, .0D, -.25D);
-                            GL11.glVertex3d(+.25D, .0D, +.25D);
-                            GL11.glVertex3d(+.50D, .0D, +.25D);
-                            GL11.glVertex3d(+.25D, .0D, -.25D);
-                        }
-                        case 5 -> {
-                            GL11.glVertex3d(+.50D, .0D, +.50D);
-                            GL11.glVertex3d(+.25D, .0D, +.25D);
-                            GL11.glVertex3d(+.50D, .0D, +.25D);
-                            GL11.glVertex3d(+.25D, .0D, +.50D);
-                            GL11.glVertex3d(+.50D, .0D, -.50D);
-                            GL11.glVertex3d(+.25D, .0D, -.25D);
-                            GL11.glVertex3d(+.50D, .0D, -.25D);
-                            GL11.glVertex3d(+.25D, .0D, -.50D);
-                            GL11.glVertex3d(-.50D, .0D, +.50D);
-                            GL11.glVertex3d(-.25D, .0D, +.25D);
-                            GL11.glVertex3d(-.50D, .0D, +.25D);
-                            GL11.glVertex3d(-.25D, .0D, +.50D);
-                            GL11.glVertex3d(-.50D, .0D, -.50D);
-                            GL11.glVertex3d(-.25D, .0D, -.25D);
-                            GL11.glVertex3d(-.50D, .0D, -.25D);
-                            GL11.glVertex3d(-.25D, .0D, -.50D);
-                        }
-                    }
-                }
-            }
-        }
-        GL11.glEnd();
-        // draw turning indicator
-        Function<ForgeDirection, Transformation[]> getTransform = (ForgeDirection direction) -> {
-            try {
-                if (direction.ordinal() == tSideHit) return new Transformation[] { ROTATION_MARKER_TRANSFORM_CENTER };
-                else if (direction.getOpposite()
-                    .ordinal() == tSideHit) {
-                        return ROTATION_MARKER_TRANSFORMS_CORNER;
-                    } else {
-                        return new Transformation[] {
-                            ROTATION_MARKER_TRANSFORMS_SIDES_TRANSFORMS[ROTATION_MARKER_TRANSFORMS_SIDES[tSideHit * 6
-                                + direction.ordinal()]] };
-                    }
-            } catch (ArrayIndexOutOfBoundsException e) {
-                return new Transformation[] {};
-            }
-
-        };
-
-        if (aIsWrench && tTile instanceof IAlignmentProvider) {
-            final IAlignment tAlignment = ((IAlignmentProvider) (tTile)).getAlignment();
-            if (tAlignment != null) {
-                for (var transform : getTransform.apply(tAlignment.getDirection())) {
-                    drawExtendedRotationMarker(transform, aIsSneaking, tAlignment);
-                }
-            }
-        }
-        if (aIsWrench && tTile instanceof IOrientable orientable
-            && !(tTile instanceof TileInterface)
-            && orientable.canBeRotated()) {
-            for (var transform : getTransform.apply(aIsSneaking ? orientable.getForward() : orientable.getUp())) {
-                drawExtendedRotationMarker(transform, aIsSneaking, orientable);
-            }
-        }
-        GL20.glUseProgram(program); // resume shader
-        GL11.glPopMatrix(); // get back to player center
-    }
-
-    private static void drawExtendedRotationMarker(Transformation transform, boolean sneaking, IAlignment alignment) {
-        if (sneaking) {
-            if (alignment.isFlipChangeAllowed()) {
-                drawFlipMarker(transform);
-            }
-        } else {
-            if (alignment.isRotationChangeAllowed()) {
-                drawRotationMarker(transform);
-            }
-        }
-    }
-
-    private static void drawExtendedRotationMarker(Transformation transform, boolean sneaking, IOrientable orientable) {
-        drawRotationMarker(transform);
-    }
-
-    private static void drawRotationMarker(Transformation transform) {
-        if (!rotationMarkerDisplayListCompiled) {
-            rotationMarkerDisplayList = GLAllocation.generateDisplayLists(1);
-            compileRotationMarkerDisplayList(rotationMarkerDisplayList);
-            rotationMarkerDisplayListCompiled = true;
-        }
-        GL11.glPushMatrix();
-        transform.glApply();
-        GL11.glCallList(rotationMarkerDisplayList);
-        GL11.glPopMatrix();
-    }
-
-    private static void compileRotationMarkerDisplayList(int displayList) {
-        GL11.glNewList(displayList, GL11.GL_COMPILE);
-        GL11.glBegin(GL_LINE_LOOP);
-        for (int i = 0; i <= ROTATION_MARKER_RESOLUTION; i++) {
-            GL11.glVertex3d(
-                Math.cos(i * Math.PI * 1.75 / ROTATION_MARKER_RESOLUTION) * 0.4,
-                0,
-                Math.sin(i * Math.PI * 1.75 / ROTATION_MARKER_RESOLUTION) * 0.4);
-        }
-        for (int i = ROTATION_MARKER_RESOLUTION; i >= 0; i--) {
-            GL11.glVertex3d(
-                Math.cos(i * Math.PI * 1.75 / ROTATION_MARKER_RESOLUTION) * 0.24,
-                0,
-                Math.sin(i * Math.PI * 1.75 / ROTATION_MARKER_RESOLUTION) * 0.24);
-        }
-        GL11.glVertex3d(0.141114561800, 0, 0);
-        GL11.glVertex3d(0.32, 0, -0.178885438199);
-        GL11.glVertex3d(0.498885438199, 0, 0);
-        GL11.glEnd();
-        GL11.glEndList();
-    }
-
-    private static void drawFlipMarker(Transformation transform) {
-        GL11.glPushMatrix();
-        transform.glApply();
-        final Tessellator t = Tessellator.instance;
-        // right shape
-        GL11.glLineStipple(4, (short) 0xAAAA);
-        GL11.glEnable(GL11.GL_LINE_STIPPLE);
-        t.startDrawing(GL11.GL_LINE_STRIP);
-        t.addVertex(0.1d, 0d, 0.04d);
-        t.addVertex(0.1d, 0d, 0.2d);
-        t.addVertex(0.35d, 0d, 0.35d);
-        t.addVertex(0.35d, 0d, -0.35d);
-        t.addVertex(0.1d, 0d, -0.2d);
-        t.addVertex(0.1d, 0d, -0.04d);
-        t.draw();
-        GL11.glDisable(GL11.GL_LINE_STIPPLE);
-        // left shape
-        t.startDrawing(GL11.GL_LINE_STRIP);
-        t.addVertex(-0.1d, 0d, 0.04d);
-        t.addVertex(-0.1d, 0d, 0.2d);
-        t.addVertex(-0.35d, 0d, 0.35d);
-        t.addVertex(-0.35d, 0d, -0.35d);
-        t.addVertex(-0.1d, 0d, -0.2d);
-        t.addVertex(-0.1d, 0d, -0.04d);
-        t.draw();
-        // arrow
-        t.startDrawing(GL11.GL_LINE_LOOP);
-        t.addVertex(0.15d, 0d, -0.04d);
-        t.addVertex(0.15d, 0d, -0.1d);
-        t.addVertex(0.25d, 0d, 0.d);
-        t.addVertex(0.15d, 0d, 0.1d);
-        t.addVertex(0.15d, 0d, 0.04d);
-        t.addVertex(-0.15d, 0d, 0.04d);
-        t.addVertex(-0.15d, 0d, 0.1d);
-        t.addVertex(-0.25d, 0d, 0.d);
-        t.addVertex(-0.15d, 0d, -0.1d);
-        t.addVertex(-0.15d, 0d, -0.04d);
-        t.draw();
-        GL11.glPopMatrix();
-    }
-
     @Override
     public boolean isClientSide() {
         return true;
@@ -568,25 +249,13 @@ public class GTClient extends GTProxy {
     @Override
     public void onPreLoad() {
         super.onPreLoad();
-
         SoundSystemConfig.setNumberNormalChannels(Client.preference.maxNumSounds);
-
         MinecraftForge.EVENT_BUS.register(new ExtraIcons());
         Minecraft.getMinecraft()
             .getResourcePackRepository().rprMetadataSerializer
                 .registerMetadataSectionType(new ColorsMetadataSectionSerializer(), ColorsMetadataSection.class);
-
-        MinecraftForge.EVENT_BUS.register(new MTEAdvDebugStructureWriter.EventHandler());
-
         new Thread(new GTCapesLoader(), "GT Cape Loader").start();
-
-        MinecraftForge.EVENT_BUS.register(mPollutionRenderer);
-        FMLCommonHandler.instance()
-            .bus()
-            .register(mPollutionRenderer);
-
         mPreference = new GTClientPreference();
-
         Materials.initClient();
     }
 
@@ -612,7 +281,13 @@ public class GTClient extends GTProxy {
         new DataStickRenderer();
         new InfiniteSprayCanRenderer();
         MinecraftForge.EVENT_BUS.register(new NEIGTConfig());
+        MinecraftForge.EVENT_BUS.register(mPollutionRenderer);
+        FMLCommonHandler.instance()
+            .bus()
+            .register(mPollutionRenderer);
         MinecraftForge.EVENT_BUS.register(new GTMouseEventHandler());
+        MinecraftForge.EVENT_BUS.register(new BlockOverlayRenderer());
+        MinecraftForge.EVENT_BUS.register(new MTEAdvDebugStructureWriter.EventHandler());
         SprayColorInfiniteKeybindHandler.init();
     }
 
@@ -729,71 +404,6 @@ public class GTClient extends GTProxy {
             mPreference = new GTClientPreference();
             GTPreLoad.loadClientConfig();
             if (e.isWorldRunning) GTValues.NW.sendToServer(new GTPacketClientPreference(mPreference));
-        }
-    }
-
-    @SubscribeEvent
-    public void onDrawBlockHighlight(DrawBlockHighlightEvent aEvent) {
-        final Block aBlock = aEvent.player.worldObj
-            .getBlock(aEvent.target.blockX, aEvent.target.blockY, aEvent.target.blockZ);
-        final TileEntity aTileEntity = aEvent.player.worldObj
-            .getTileEntity(aEvent.target.blockX, aEvent.target.blockY, aEvent.target.blockZ);
-
-        if (GTUtility.isStackInList(aEvent.currentItem, GregTechAPI.sWrenchList)) {
-            if (aTileEntity instanceof ITurnable || ROTATABLE_VANILLA_BLOCKS.contains(aBlock)
-                || aTileEntity instanceof IWrenchable
-                || (aTileEntity instanceof IOrientable orientable && orientable.canBeRotated())
-                || (aBlock instanceof BlockFrameBox)) drawGrid(aEvent, false, true, aEvent.player.isSneaking());
-            return;
-        }
-
-        // If there is no tile entity and the block is a frame box block, still draw the grid if a cover is held
-        if (aTileEntity == null && aBlock instanceof BlockFrameBox) {
-            if (CoverRegistry.isCover(aEvent.currentItem)) {
-                drawGrid(aEvent, true, false, aEvent.player.isSneaking());
-            }
-            return;
-        }
-
-        if (!(aTileEntity instanceof ICoverable)) return;
-
-        if (aEvent.player.isSneaking() && aTileEntity instanceof IGregTechTileEntity gtEntity
-            && gtEntity.getMetaTileEntity() instanceof MetaPipeEntity) {
-            if (aEvent.currentItem != null && aEvent.currentItem.getItem() instanceof ItemMachines
-                && GregTechAPI.METATILEENTITIES[aEvent.currentItem.getItemDamage()] instanceof MetaPipeEntity) {
-                drawGrid(aEvent, false, false, false);
-            }
-        }
-
-        if (GTUtility.isStackInList(aEvent.currentItem, GregTechAPI.sWireCutterList)
-            || GTUtility.isStackInList(aEvent.currentItem, GregTechAPI.sSolderingToolList)
-                && aEvent.player.isSneaking()) {
-            if (!((ICoverable) aTileEntity).hasCoverAtSide(ForgeDirection.getOrientation(aEvent.target.sideHit)))
-                drawGrid(aEvent, false, false, aEvent.player.isSneaking());
-            return;
-        }
-
-        if ((aEvent.currentItem == null && aEvent.player.isSneaking())
-            || GTUtility.isStackInList(aEvent.currentItem, GregTechAPI.sCrowbarList)
-            || GTUtility.isStackInList(aEvent.currentItem, GregTechAPI.sScrewdriverList)) {
-            if (!((ICoverable) aTileEntity).hasCoverAtSide(ForgeDirection.getOrientation(aEvent.target.sideHit)))
-                for (final ForgeDirection tSide : ForgeDirection.VALID_DIRECTIONS) {
-                    if (((ICoverable) aTileEntity).hasCoverAtSide(tSide)) {
-                        drawGrid(aEvent, true, false, true);
-                        return;
-                    }
-                }
-            return;
-        }
-
-        if (CoverRegistry.isCover(aEvent.currentItem)) {
-            if (!((ICoverable) aTileEntity).hasCoverAtSide(ForgeDirection.getOrientation(aEvent.target.sideHit)))
-                drawGrid(aEvent, true, false, aEvent.player.isSneaking());
-        }
-
-        if (GTUtility.areStacksEqual(ItemList.Tool_Cover_Copy_Paste.get(1), aEvent.currentItem, true)) {
-            if (!((ICoverable) aTileEntity).hasCoverAtSide(ForgeDirection.getOrientation(aEvent.target.sideHit)))
-                drawGrid(aEvent, true, false, aEvent.player.isSneaking());
         }
     }
